test: metrics subscription testing support
